make JsonExtractorInterface:extractFromJson(): array
