+++ src/Domain/Crawler/Site/Barrons/BarronsSiteCrawler.php
@@ -7,7 +7,6 @@ namespace Chemaclass\TickerNews\Domain\Crawler\Site\Barrons;
 use Chemaclass\TickerNews\Domain\Crawler\SiteCrawlerInterface;
 use Chemaclass\TickerNews\Domain\ReadModel\Site;
 use Chemaclass\TickerNews\Domain\ReadModel\Ticker;
-use DOMNode;
 use Symfony\Contracts\HttpClient\HttpClientInterface;
 
 final class BarronsSiteCrawler implements SiteCrawlerInterface
@@ -26,18 +25,17 @@ final class BarronsSiteCrawler implements SiteCrawlerInterface
 
     public function crawl(HttpClientInterface $httpClient, Ticker $ticker): Site
     {
-        $symbol = strtolower($ticker->symbol());
+        $symbol = mb_strtolower($ticker->symbol());
 
-        $url = sprintf(self::REQUEST_URL, strtolower($symbol));
+        $url = sprintf(self::REQUEST_URL, mb_strtolower($symbol));
 
         $html = $httpClient
             ->request(self::REQUEST_METHOD, $url, [
                 'headers' => [
                     'User-Agent' => 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/74.0.3729.169 Safari/537.36',
-                ]
+                ],
             ])
             ->getContent();
-
 
         $crawled = [];
