--- src/Domain/Crawler/Site/Barrons/HtmlExtractor/CompanyName.php
@@ -1,12 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-
-namespace Chemaclass\TickerNews\Domain\Crawler\Site\Barrons\HtmlExtractor;
-
-
-final class CompanyName implements HtmlExtractorInterface
-{
-
-}
