+++ src/Domain/Crawler/Site/FinanceYahoo/JsonExtractor/QuoteSummaryStore/RegularMarketPrice.php
@@ -8,10 +8,10 @@ use Chemaclass\TickerNews\Domain\Crawler\Site\FinanceYahoo\JsonExtractorInterfac
 
 final class RegularMarketPrice implements JsonExtractorInterface
 {
-    public function extractFromJson(array $json): string
+    public function extractFromJson(array $json): array
     {
         $quoteSummaryStore = $json['context']['dispatcher']['stores']['QuoteSummaryStore'];
 
-        return  (string) $quoteSummaryStore['price']['regularMarketPrice']['raw'];
+        return $quoteSummaryStore['price']['regularMarketPrice'];
     }
 }
