+++ src/Domain/Notifier/Policy/Condition/FoundMoreNews.php
@@ -33,7 +33,7 @@ final class FoundMoreNews implements PolicyConditionInterface
     private function findLatestPubtimeFromNews(Company $company): string
     {
         $reduced = array_reduce(
-            (array) $company->info('news')->get(),
+            (array) $company->info('news'),
             static function (?array $carry, array $current): array {
                 if (null === $carry) {
                     return $current;
